server/pyproject.toml
[project]
name = "os-backend"
version = "0.1.0"
description = "Axon API server"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic-settings>=2.7.0",
    "python-jose[cryptography]>=3.3.0",
    "eth-account>=0.13.0",
    "httpx>=0.28.0",
    "psycopg2-binary>=2.9.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0",
]

[tool.setuptools.packages.find]
include = ["axon_server*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
